Reader Edit Hourly rate possible

+++ PerfectSelfApp/Sources/Common/Global.swift
@@ -31,7 +31,7 @@ struct ReaderProfileDetail: Codable {
     let uid: String
     let userName: String
     let title: String
-    let hourlyPrice: Float
+    let hourlyPrice: Int
     let others: Int
     let voiceType: Int
     let about: String
